feat: add basic general settings
