+++ pages/godam/index.scss
@@ -1,119 +1,144 @@
 #easydam-settings {
 
-    .easydam-settings {
-    
-        &-header {
-            background: #fff;
-        }
-    
-        &-shadow {
-            box-shadow: 0 1px 4px rgba(18,25,97,.0779552);
-        }
-    
-        &-label {
-            display: block;
-            font-size: 13px;
-            font-weight: 500;
-            line-height: 1.4;
-            margin-bottom: calc(4px* 2);
-            padding: 0;
-            text-transform: none;
-        }
-    
-        &-help {
-            margin-top: calc(4px* 2);
-            margin-bottom: 0;
-            font-size: 12px;
-            font-style: normal;
-            color: #757575;
-        }
-    }
+	.easydam-settings {
 
-    .invalid-license-key {
-        input {
-            border-color: #fb2c36;
-        }
-    }
-    
-    input[type="number"] {
-        max-width: 100px;
-    }
-    
-    .sidebar-nav-item {
-        @apply flex items-center gap-2 p-2 font-medium rounded-md text-gray-600 hover:text-gray-900 border border-transparent transition-all;
-    
-        &.active {
-            @apply bg-white font-semibold focus:ring-2 ring-indigo-200 text-gray-900 drop-shadow-sm;
-            border: 1px solid #ddd;
-        }
-    }
-    
-    
-    .components-panel {
-        box-shadow: 0 1px 4px rgba(18,25,97,.0779552);
-        border-radius: 0.25rem;
-    }
-    
-    .components-panel__header {
-        font-size: 14px;
-    }
-    
-    .components-base-control__label {
-        font-size: 13px !important;
-        text-transform: none !important;
-    }
+		&-header {
+			background: #fff;
+		}
 
-    // Circular progress bar
-    .circle-container {
-        position: relative;
-        width: 80px;
-        height: 80px;
+		&-shadow {
+			box-shadow: 0 1px 4px rgba(18, 25, 97, .0779552);
+		}
 
-        .circle {
-            width: 80px;
-            height: 80px;
-            border-radius: 50%;
-            background: conic-gradient(var(--wp-components-color-accent,var(--wp-admin-theme-color,#3858e9)) var(--percentage), #e0e0e0 0);
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            font-size: 20px;
-            font-weight: bold;
-            color: #333;
+		&-label {
+			display: block;
+			font-size: 13px;
+			font-weight: 500;
+			line-height: 1.4;
+			margin-bottom: calc(4px* 2);
+			padding: 0;
+			text-transform: none;
+		}
 
-            &.red {
-                background: conic-gradient(#d14949 var(--percentage), #e0e0e0 0);
-            }
-        }
+		&-help {
+			margin-top: calc(4px* 2);
+			margin-bottom: 0;
+			font-size: 12px;
+			font-style: normal;
+			color: #757575;
+		}
+	}
 
-        .data {
-            position: absolute;
-            top: 50%;
-            left: 50%;
-            width: 65px;
-            height: 65px;
-            border-radius: 1000px;
-            transform: translate(-50%, -50%);
-            background: #fff;
-            display: flex;
-            justify-content: center;
-            align-items: center;         
-        }
-    }
+	.invalid-license-key {
+		input {
+			border-color: #fb2c36;
+		}
+	}
 
-    .premium-feature-overlay {
-        position: absolute;
-        inset: 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        z-index: 99;
-        background-color: rgb(255, 255, 255, 0.4);
-        border-radius: 0.25rem;
+	input[type="number"] {
+		max-width: 100px;
+	}
 
-        button {
-            box-shadow: 0 0 10px var(--wp-components-color-accent, #ab3a6d49);
-        }
-    }
+	.sidebar-nav-item {
+		@apply flex items-center gap-2 p-2 font-medium rounded-md text-gray-600 hover:text-gray-900 border border-transparent transition-all;
+
+		&.active {
+			@apply bg-white font-semibold focus:ring-2 ring-indigo-200 text-gray-900 drop-shadow-sm;
+			border: 1px solid #ddd;
+		}
+	}
+
+
+	// Circular progress bar
+	.circle-container {
+		position: relative;
+		width: 80px;
+		height: 80px;
+
+		.circle {
+			width: 80px;
+			height: 80px;
+			border-radius: 50%;
+			background: conic-gradient(var(--wp-components-color-accent, var(--wp-admin-theme-color, #3858e9)) var(--percentage), #e0e0e0 0);
+			display: flex;
+			align-items: center;
+			justify-content: center;
+			font-size: 20px;
+			font-weight: bold;
+			color: #333;
+
+			&.red {
+				background: conic-gradient(#d14949 var(--percentage), #e0e0e0 0);
+			}
+		}
+
+		.data {
+			position: absolute;
+			top: 50%;
+			left: 50%;
+			width: 65px;
+			height: 65px;
+			border-radius: 1000px;
+			transform: translate(-50%, -50%);
+			background: #fff;
+			display: flex;
+			justify-content: center;
+			align-items: center;
+		}
+	}
+
+	.premium-feature-overlay {
+		position: absolute;
+		inset: 0;
+		display: flex;
+		justify-content: center;
+		align-items: center;
+		z-index: 99;
+		background-color: rgb(255, 255, 255, 0.4);
+		border-radius: 0.25rem;
+
+		button {
+			box-shadow: 0 0 10px var(--wp-components-color-accent, #ab3a6d49);
+		}
+	}
+
+	.easydam-settings__container {
+
+		max-width: 80rem;
+		margin: 2rem auto;
+
+		&__tabs {
+
+			margin-bottom: 2rem;
+
+			nav {
+				display: flex;
+				gap: 1.5rem;
+
+				border-bottom: 2px solid rgba(228, 228, 228, 1);
+
+				a {
+
+					font-size: 1rem;
+					line-height: 1.375rem;
+					font-weight: 500;
+					color: rgba(119, 119, 119, 1);
+
+					padding-bottom: 1rem;
+
+					&.active {
+						color: rgba(33, 33, 33, 1);
+						border-bottom: 2px solid;
+						border-image-slice: 1;
+						border-image-source: linear-gradient(83.85deg, #AB3A6C -9.3%, #E6533A 120.31%);	
+					}
+
+					&:focus {
+						box-shadow: none;
+					}
+				}
+			}
+		}
+	}
+
 }
-
